chore: organize styles
